requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httpx==0.27.2
python-multipart==0.0.12
jinja2==3.1.4
pydantic==2.9.2
